+++ src/elements/processors/AssetProcessor.php
@@ -3,11 +3,11 @@
 namespace venveo\bulkedit\elements\processors;
 
 use craft\elements\Asset;
-use craft\elements\User;
 use craft\helpers\ArrayHelper;
 use craft\records\FieldLayout;
-use craft\services\Users;
+use craft\web\User;
 use venveo\bulkedit\base\AbstractElementTypeProcessor;
+use venveo\bulkedit\Plugin;
 
 class AssetProcessor extends AbstractElementTypeProcessor
 {
@@ -40,5 +40,16 @@ class AssetProcessor extends AbstractElementTypeProcessor
     public static function getType(): string
     {
         return get_class(new Asset);
+    }
+
+    /**
+     * Return whether a given user has permission to perform bulk edit actions on these elements
+     * @param $elementIds
+     * @param $user
+     * @return bool
+     */
+    public static function hasPermission($elementIds, User $user): bool
+    {
+        return $user->checkPermission(Plugin::PERMISSION_BULKEDIT_ASSETS);
     }
 }