+++ src/migrations/Install.php
@@ -2,10 +2,7 @@
 
 namespace venveo\bulkedit\migrations;
 
-use venveo\bulkedit\Plugin;
-
 use Craft;
-use craft\config\DbConfig;
 use craft\db\Migration;
 
 /**
@@ -59,25 +56,26 @@ class Install extends Migration
             $this->createTable(
                 '{{%bulkedit_editcontext}}',
                 [
-                    'id'          => $this->primaryKey(),
+                    'id' => $this->primaryKey(),
                     'dateCreated' => $this->dateTime()->notNull(),
                     'dateUpdated' => $this->dateTime()->notNull(),
-                    'uid'         => $this->uid(),
+                    'uid' => $this->uid(),
 
                     'siteId' => $this->integer()->notNull(),
                     'ownerId' => $this->integer()->notNull(),
                     'elementIds' => $this->string(1024)->notNull(),
                     'fieldIds' => $this->string(1024)->notNull(),
+                    'elementType' => $this->string()->notNull(),
                 ]
             );
             $this->createTable(
                 '{{%bulkedit_history}}',
                 [
-                    'id'          => $this->primaryKey(),
+                    'id' => $this->primaryKey(),
                     'dateCreated' => $this->dateTime()->notNull(),
                     'dateUpdated' => $this->dateTime()->notNull(),
                     'status' => $this->string()->notNull()->defaultValue('pending'),
-                    'uid'         => $this->uid(),
+                    'uid' => $this->uid(),
 
                     'contextId' => $this->integer()->notNull(),
                     'elementId' => $this->integer()->notNull(),