+++ src/migrations/m181117_192854_increase_data_column_size.php
@@ -2,7 +2,6 @@
 
 namespace venveo\bulkedit\migrations;
 
-use Craft;
 use craft\db\Migration;
 
 /**