+++ src/queue/jobs/SaveBulkEditJob.php
@@ -11,9 +11,11 @@
 namespace venveo\bulkedit\queue\jobs;
 
 use Craft;
+use craft\base\ElementInterface;
 use craft\queue\BaseJob;
 use venveo\bulkedit\Plugin;
 use venveo\bulkedit\records\EditContext;
+use venveo\bulkedit\records\History;
 use yii\base\Exception;
 
 /**
@@ -40,36 +42,47 @@ class SaveBulkEditJob extends BaseJob
      */
     public function execute($queue = null)
     {
-        $elementIds = Plugin::$plugin->bulkEdit->getPendingElementIdsFromContext($this->context);
-        $totalSteps = count($elementIds);
+        $elementHistories = Plugin::$plugin->bulkEdit->getPendingElementsHistoriesFromContext($this->context);
+        $totalSteps = $elementHistories->count();
         try {
-            foreach ($elementIds as $key => $elementId) {
+            $currentRow = 0;
+            /**
+             * @var History $elementHistory */
+            foreach ($elementHistories->each() as $elementHistory) {
+                $elementId = $elementHistory->elementId;
+                /** @var ElementInterface $element */
                 $element = Craft::$app->getElements()->getElementById($elementId, null, $this->context->siteId);
                 if (!$element) {
+                    Craft::warning('Could not locate an element in a bulk save job: ' . $elementId, __METHOD__);
                     continue;
                 }
+
+                if (get_class($element) !== $this->context->elementType) {
+                    throw new \Exception('Unexpected element type encountered!');
+                }
+
                 $history = Plugin::$plugin->bulkEdit->getPendingHistoryForElement($this->context, $element->id)->all();
                 try {
                     Craft::info('Starting processing bulk edit job', __METHOD__);
                     Plugin::$plugin->bulkEdit->processHistoryItemsForElement($history, $element);
                 } catch (\Exception $e) {
-                    Craft::error('Could not save element in bulk edit job... '. $e->getMessage(), __METHOD__);
+                    Craft::error('Could not save element in bulk edit job... ' . $e->getMessage(), __METHOD__);
                     throw new Exception('Couldn’t save element ' . $element->id . ' (' . get_class($element) . ')');
                 } catch (\Throwable $e) {
                     throw $e;
                 }
-
-                if (($key + 1) === $totalSteps) {
+                if (($currentRow + 1) === (int)$totalSteps) {
                     try {
                         $this->context->delete();
                     } catch (\Exception $e) {
                         throw new Exception('Couldn’t delete context: ' . $e->getMessage());
                     }
                 }
-                $this->setProgress($queue, ($key + 1) / $totalSteps, 'Element '. ($key + 1) . ' of '. $totalSteps);
+                $this->setProgress($queue, ($currentRow + 1) / $totalSteps, 'Element ' . ($currentRow + 1) . ' of ' . $totalSteps);
+                $currentRow++;
             }
         } catch (\Exception $e) {
-            Craft::error('Failed to save... '. $e->getMessage(), __METHOD__);
+            Craft::error('Failed to save... ' . $e->getMessage(), __METHOD__);
             throw $e;
         }
     }